+++ DatadogInternal/Sources/NetworkInstrumentation/W3C/W3CHTTPHeadersReader.swift
@@ -33,4 +33,15 @@ public class W3CHTTPHeadersReader: TracePropagationHeadersReader {
             parentSpanID: nil
         )
     }
+
+    public var sampled: Bool? {
+        if let traceparent = httpHeaderFields[W3CHTTPHeaders.traceparent] {
+            guard let sampled = traceparent.components(separatedBy: W3CHTTPHeaders.Constants.separator).last else {
+                return nil
+            }
+            return sampled == W3CHTTPHeaders.Constants.sampledValue
+        }
+
+        return nil
+    }
 }